Rename NaiveTSP to HeldKarpTSP and add new tests and docs.
